Multimodule - create dao, facade and service modules
