+++ formula1-dao/src/main/java/cz/muni/pa165/teamwhite/formula1/persistence/enums/ComponentType.java
@@ -1,0 +1,12 @@
+package cz.muni.pa165.teamwhite.formula1.persistence.enums;
+
+/**
+ * @author Jiří Andrlík
+ */
+public enum ComponentType {
+    ENGINE,
+    SUSPENSION,
+    TRANSMISSION,
+    RIMS,
+    REARSPOILER
+}
