+++ formula1-dao/src/main/java/cz/muni/pa165/teamwhite/formula1/persistence/enums/Role.java
@@ -1,0 +1,9 @@
+package cz.muni.pa165.teamwhite.formula1.persistence.enums;
+
+/**
+ * @author Karolina Hecova
+ */
+public enum Role {
+    MANAGER,
+    ENGINEER
+}
