--- src/main/java/cz/muni/pa165/teamwhite/formula1/enums/ComponentType.java
@@ -1,12 +1,0 @@
-package cz.muni.pa165.teamwhite.formula1.enums;
-
-/**
- * @author Jiří Andrlík
- */
-public enum ComponentType {
-    ENGINE,
-    SUSPENSION,
-    TRANSMISSION,
-    RIMS,
-    REARSPOILER
-}
