--- src/main/java/cz/muni/pa165/teamwhite/formula1/enums/Role.java
@@ -1,9 +1,0 @@
-package cz.muni.pa165.teamwhite.formula1.enums;
-
-/**
- * @author Karolina Hecova
- */
-public enum Role {
-    MANAGER,
-    ENGINEER
-}
